Created new module about queries
