+++ User.py
@@ -1,0 +1,6 @@
+import graphene
+
+class User(graphene.ObjectType):
+    id = graphene.ID()
+    username = graphene.String()
+    lastLogin = graphene.Date()
